+++ datajunction-server/datajunction_server/api/graphql/queries/nodes.py
@@ -229,13 +229,6 @@ async def find_nodes_paginated(
             description="Filter to nodes in this namespace",
         ),
     ] = None,
-    recursive: Annotated[
-        bool,
-        strawberry.argument(
-            description="When false, return only nodes whose namespace equals "
-            "`namespace` exactly (no descendants). No effect when namespace is null.",
-        ),
-    ] = True,
     mode: Annotated[
         NodeMode | None,
         strawberry.argument(
@@ -310,7 +303,6 @@ async def find_nodes_paginated(
         dimensions=dimensions,
         edited_by=edited_by,
         namespace=namespace,
-        recursive=recursive,
         limit=limit + 1,
         before=before,
         after=after,
@@ -339,7 +331,6 @@ async def find_nodes_paginated(
             dimensions=dimensions,
             edited_by=edited_by,
             namespace=namespace,
-            recursive=recursive,
             mode=mode,
             owned_by=owned_by,
             include_team=include_team,
