+++ datajunction-server/datajunction_server/api/graphql/resolvers/nodes.py
@@ -214,7 +214,6 @@ async def find_nodes_by(
     tags: Optional[List[str]] = None,
     edited_by: Optional[str] = None,
     namespace: Optional[str] = None,
-    recursive: bool = True,
     limit: Optional[int] = 100,
     before: Optional[str] = None,
     after: Optional[str] = None,
@@ -276,7 +275,6 @@ async def find_nodes_by(
             tags,
             edited_by,
             namespace,
-            recursive,
             limit,
             before,
             after,
@@ -317,7 +315,6 @@ async def count_nodes_by(
     tags: Optional[List[str]] = None,
     edited_by: Optional[str] = None,
     namespace: Optional[str] = None,
-    recursive: bool = True,
     mode: Optional[NodeMode] = None,
     owned_by: Optional[str] = None,
     include_team: bool = False,
@@ -352,7 +349,6 @@ async def count_nodes_by(
             tags=tags,
             edited_by=edited_by,
             namespace=namespace,
-            recursive=recursive,
             mode=mode,
             owned_by=owned_by_list,
             missing_description=missing_description,
