+++ datajunction-server/datajunction_server/api/graphql/schema.graphql
@@ -549,11 +549,6 @@ type Query {
     """Filter to nodes in this namespace"""
     namespace: String = null
 
-    """
-    When false, return only nodes whose namespace equals `namespace` exactly (no descendants). No effect when namespace is null.
-    """
-    recursive: Boolean! = true
-
     """Filter to nodes with this mode (published or draft)"""
     mode: NodeMode = null
