+++ datajunction-ui/src/app/pages/NamespacePage/__tests__/namespaceOptions.test.js
@@ -3,7 +3,6 @@ import {
   buildNamespaceOptions,
   findHierarchyNode,
   immediateChildren,
-  buildJumpTree,
   searchNamespaces,
 } from '../namespaceOptions';
 
@@ -108,38 +107,5 @@ describe('immediateChildren', () => {
   it('returns [] for a leaf or missing path', () => {
     expect(immediateChildren(jumpHierarchy, 'growth.experiments')).toEqual([]);
     expect(immediateChildren(jumpHierarchy, 'nope')).toEqual([]);
-  });
-});
-
-describe('buildJumpTree', () => {
-  it('shows siblings at each level, current collapsed, no current children', () => {
-    const rows = buildJumpTree(jumpHierarchy, 'growth');
-    expect(rows.map(r => r.path)).toEqual([
-      'default',
-      'finance',
-      'growth',
-      'marketing',
-    ]);
-    const growth = rows.find(r => r.path === 'growth');
-    expect(growth.isCurrent).toBe(true);
-    expect(growth.depth).toBe(0);
-  });
-  it('descends along the current path, excluding current children', () => {
-    const rows = buildJumpTree(jumpHierarchy, 'growth.metrics');
-    expect(rows.map(r => r.path)).toEqual([
-      'default',
-      'finance',
-      'growth',
-      'growth.experiments',
-      'growth.metrics',
-      'marketing',
-    ]);
-    expect(rows.find(r => r.path === 'growth').isAncestor).toBe(true);
-    expect(rows.find(r => r.path === 'growth.metrics').isCurrent).toBe(true);
-    // growth.metrics.dau (a child of current) is NOT present
-    expect(rows.find(r => r.path === 'growth.metrics.dau')).toBeUndefined();
-  });
-  it('returns [] when no currentPath', () => {
-    expect(buildJumpTree(jumpHierarchy, '')).toEqual([]);
   });
 });
