+++ datajunction-ui/src/app/services/DJService.js
@@ -73,12 +73,11 @@ export const DataJunctionAPI = {
       missingDescription = false,
       hasMaterialization = false,
       orphanedDimension = false,
-      recursive = true,
       search = null,
     } = {},
   ) {
     const query = `
-      query ListNodes($namespace: String, $nodeTypes: [NodeType!], $tags: [String!], $editedBy: String, $mode: NodeMode, $before: String, $after: String, $limit: Int, $orderBy: NodeSortField, $ascending: Boolean, $ownedBy: String, $statuses: [NodeStatus!], $missingDescription: Boolean, $hasMaterialization: Boolean, $orphanedDimension: Boolean, $recursive: Boolean, $search: String) {
+      query ListNodes($namespace: String, $nodeTypes: [NodeType!], $tags: [String!], $editedBy: String, $mode: NodeMode, $before: String, $after: String, $limit: Int, $orderBy: NodeSortField, $ascending: Boolean, $ownedBy: String, $statuses: [NodeStatus!], $missingDescription: Boolean, $hasMaterialization: Boolean, $orphanedDimension: Boolean, $search: String) {
         findNodesPaginated(
           namespace: $namespace
           nodeTypes: $nodeTypes
@@ -95,7 +94,6 @@ export const DataJunctionAPI = {
           missingDescription: $missingDescription
           hasMaterialization: $hasMaterialization
           orphanedDimension: $orphanedDimension
-          recursive: $recursive
           search: $search
         ) {
           pageInfo {
@@ -167,7 +165,6 @@ export const DataJunctionAPI = {
             missingDescription: missingDescription,
             hasMaterialization: hasMaterialization,
             orphanedDimension: orphanedDimension,
-            recursive: recursive,
             search: search,
           },
         }),
